requirements.txt
# Requirements for SeqXGPT + BERT AI Text Detection Pipeline

# Core dependencies
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# HuggingFace
transformers>=4.30.0
datasets>=2.14.0
tokenizers>=0.13.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Evaluation metrics
tabulate>=0.9.0

# Optional: For paraphrasing/translation attacks
sentencepiece>=0.1.99
sacremoses>=0.0.53

# Development
jupyter>=1.0.0
ipython>=8.0.0
